+++ Form/RegistrationFormHandlerInterface.php
@@ -12,7 +12,7 @@
 namespace HWI\Bundle\OAuthBundle\Form;
 
 use HWI\Bundle\OAuthBundle\OAuth\Response\UserResponseInterface;
-use Symfony\Component\Form\Form;
+use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
@@ -27,11 +27,7 @@ interface RegistrationFormHandlerInterface
     /**
      * Processes the form for a given request.
      *
-     * @param Request               $request         Active request
-     * @param Form                  $form            Form to process
-     * @param UserResponseInterface $userInformation OAuth response
-     *
      * @return bool True if the processing was successful
      */
-    public function process(Request $request, Form $form, UserResponseInterface $userInformation);
+    public function process(Request $request, FormInterface $form, UserResponseInterface $userInformation);
 }