+++ Tests/Controller/LoginControllerTest.php
@@ -117,9 +117,11 @@ class LoginControllerTest extends TestCase
 
     public function testRegistrationRedirect()
     {
-        $this->request->attributes = new ParameterBag([
-            $this->getAuthenticationErrorKey() => new AccountNotLinkedException(),
-        ]);
+        $this->request->attributes = new ParameterBag(
+            [
+                Security::AUTHENTICATION_ERROR => new AccountNotLinkedException(),
+            ]
+        );
 
         $this->mockAuthorizationCheck(false);
 
@@ -140,9 +142,11 @@ class LoginControllerTest extends TestCase
 
         $authenticationException = new AuthenticationException();
 
-        $this->request->attributes = new ParameterBag([
-            $this->getAuthenticationErrorKey() => $authenticationException,
-        ]);
+        $this->request->attributes = new ParameterBag(
+            [
+                Security::AUTHENTICATION_ERROR => $authenticationException,
+            ]
+        );
 
         $this->twig->expects($this->once())
             ->method('render')
@@ -160,7 +164,7 @@ class LoginControllerTest extends TestCase
 
         $this->session->expects($this->once())
             ->method('has')
-            ->with($this->getAuthenticationErrorKey())
+            ->with(Security::AUTHENTICATION_ERROR)
             ->willReturn(true)
         ;
 
@@ -168,7 +172,7 @@ class LoginControllerTest extends TestCase
 
         $this->session->expects($this->once())
             ->method('get')
-            ->with($this->getAuthenticationErrorKey())
+            ->with(Security::AUTHENTICATION_ERROR)
             ->willReturn($authenticationException)
         ;
 
@@ -182,7 +186,7 @@ class LoginControllerTest extends TestCase
         $controller->connectAction($this->request);
     }
 
-    private function mockAuthorizationCheck($granted = true)
+    private function mockAuthorizationCheck(bool $granted = true): void
     {
         $this->authorizationChecker->expects($this->once())
             ->method('isGranted')
@@ -191,12 +195,7 @@ class LoginControllerTest extends TestCase
         ;
     }
 
-    private function getAuthenticationErrorKey(): string
-    {
-        return Security::AUTHENTICATION_ERROR;
-    }
-
-    private function createController(bool $connect = true, string $grantRule = 'IS_AUTHENTICATED_REMEMBERED'): LoginController
+    private function createController(): LoginController
     {
         return new LoginController(
             $this->authenticationUtils,
@@ -204,8 +203,8 @@ class LoginControllerTest extends TestCase
             $this->authorizationChecker,
             $this->requestStack,
             $this->twig,
-            $connect,
-            $grantRule
+            true,
+            'IS_AUTHENTICATED_REMEMBERED'
         );
     }
 }