+++ Tests/Functional/Controller/ConnectControllerTest.php
@@ -23,22 +23,15 @@ use Psr\Http\Message\ResponseInterface;
 use Symfony\Bundle\FrameworkBundle\KernelBrowser;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\BrowserKit\Cookie;
+use Symfony\Component\HttpFoundation\Exception\SessionNotFoundException;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 
-/**
- * uses FOSUserBundle which itself contains lots of deprecations.
- *
- * @group legacy
- */
 final class ConnectControllerTest extends WebTestCase
 {
     protected function setUp(): void
     {
         static::$class = AppKernel::class;
-        if (!class_exists(\FOS\UserBundle\Model\User::class)) {
-            $this->markTestSkipped('FOSUserBundle not installed.');
-        }
     }
 
     public static function getKernelClass(): string
@@ -81,10 +74,10 @@ final class ConnectControllerTest extends WebTestCase
 
         $form = $crawler->filter('form')->form();
 
-        $form['fos_user_registration_form[email]']->setValue('test@example.com');
-        $form['fos_user_registration_form[username]']->setValue('username');
-        $form['fos_user_registration_form[plainPassword][first]']->setValue('bar');
-        $form['fos_user_registration_form[plainPassword][second]']->setValue('bar');
+        $form['registration[email]']->setValue('test@example.com');
+        $form['registration[username]']->setValue('foo');
+        $form['registration[plainPassword][first]']->setValue('bar');
+        $form['registration[plainPassword][second]']->setValue('bar');
 
         $crawler = $client->submit($form);
         $response = $client->getResponse();
@@ -123,7 +116,7 @@ final class ConnectControllerTest extends WebTestCase
         $response = $client->getResponse();
 
         $this->assertSame(200, $response->getStatusCode(), $response->getContent());
-        $this->assertSame(1, $crawler->filter('.fos_user_registration_register')->count(), $response->getContent());
+        $this->assertSame(1, $crawler->filter('.registration_register')->count(), $response->getContent());
 
         $form = $crawler->filter('form')->form();
 
@@ -160,7 +153,11 @@ final class ConnectControllerTest extends WebTestCase
 
         $session = null;
         if (method_exists($requestStack, 'getSession')) {
-            $session = $requestStack->getSession();
+            try {
+                $session = $requestStack->getSession();
+            } catch (SessionNotFoundException $e) {
+                // Ignore & fallback to service
+            }
         } elseif ((null !== $request = $requestStack->getCurrentRequest()) && $request->hasSession()) {
             $session = $request->getSession();
         }