+++ Tests/Functional/Controller/LoginControllerTest.php
@@ -15,30 +15,17 @@ namespace HWI\Bundle\OAuthBundle\Tests\Functional\Controller;
 
 use HWI\Bundle\OAuthBundle\Security\Core\Exception\AccountNotLinkedException;
 use HWI\Bundle\OAuthBundle\Tests\Fixtures\CustomOAuthToken;
-use HWI\Bundle\OAuthBundle\Tests\Functional\WebTestCase;
 use Psr\Http\Client\ClientInterface;
+use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\BrowserKit\Cookie;
+use Symfony\Component\HttpFoundation\Exception\SessionNotFoundException;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
 use Symfony\Component\Security\Core\Security;
 
-/**
- * uses FOSUserBundle which itself contains lots of deprecations.
- *
- * @group legacy
- */
 final class LoginControllerTest extends WebTestCase
 {
-    protected function setUp(): void
-    {
-        if (!class_exists(\FOS\UserBundle\Model\User::class)) {
-            $this->markTestSkipped('FOSUserBundle not installed.');
-        }
-
-        parent::setUp();
-    }
-
     public function testLoginPage(): void
     {
         $client = static::createClient();
@@ -106,7 +93,11 @@ final class LoginControllerTest extends WebTestCase
 
         $session = null;
         if (method_exists($requestStack, 'getSession')) {
-            $session = $requestStack->getSession();
+            try {
+                $session = $requestStack->getSession();
+            } catch (SessionNotFoundException $e) {
+                // Ignore & fallback to service
+            }
         } elseif ((null !== $request = $requestStack->getCurrentRequest()) && $request->hasSession()) {
             $session = $request->getSession();
         }